requirements.txt
# Core ML/AI Libraries
ultralytics>=8.0.207
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=9.4.0

# Web Framework
flask>=2.3.0
waitress>=2.1.0

# Google Drive API
google-api-python-client>=2.106.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data Science
pandas>=1.5.0
matplotlib>=3.7.0
scipy>=1.11.0

# Dataset Management
roboflow>=1.1.7

# Video Processing
yt-dlp>=2023.0.0

# Object Tracking
lapx>=0.5.0

# Utilities
tqdm>=4.64.0
python-dotenv>=1.0.0 